requirements.txt
apache-airflow==2.6.3
apache-airflow-providers-google==10.1.0
apache-airflow-providers-http==4.1.0
dvc
dvc-gdrive
requests
beautifulsoup4
pandas
